Se reorganiza el proyecto, ahora ya no se guardan audios temporales.
