+++ package.json
@@ -17,6 +17,7 @@
     "@builderbot/bot": "1.1.7",
     "@builderbot/provider-baileys": "1.1.7",
     "dotenv": "16.4.5",
+    "express": "4.19.2",
     "openai": "4.52.7",
     "sharp": "^0.33.4"
   },
